Refactoring getBookings() and getOwnerBookings() are done.
