+++ src/main/java/ru/practicum/shareit/booking/BookingServiceImpl.java
@@ -13,10 +13,8 @@ import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.user.UserRepository;
 import ru.practicum.shareit.user.model.User;
 
-import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Map;
 
 @Slf4j
 @Service
@@ -27,6 +25,9 @@ public class BookingServiceImpl implements BookingService {
     private final UserRepository userRepository;
 
     private final ItemRepository itemRepository;
+
+    @Autowired
+    private Map<String, BookingGenerator> bookingGenerators;
 
     @Autowired
     public BookingServiceImpl(BookingRepository bookingRepository, UserRepository userRepository, ItemRepository itemRepository) {
@@ -105,161 +106,23 @@ public class BookingServiceImpl implements BookingService {
                 "вещь не принадлежит");
     }
 
-
     @Override
     public List<BookingResponseDto> getBookings(String state, Long userId) {
-        User user = userRepository.findById(userId).orElseThrow(() -> new NotFoundException("Пользователя с таким id" +
-                " не существует"));
-        if (state.equals("ALL")) {
-            List<Booking> bookings = bookingRepository.findAllByBooker(user);
-            log.info("Возвращаем список всех бронирований для пользователя с id: {} {}", userId, bookings);
-            return bookings.stream()
-                    .sorted(this::compareAll)
-                    .map(BookingConverter::toBookingResponseDto)
-                    .collect(Collectors.toList());
+        BookingGenerator bookingUserGenerator = bookingGenerators.get(state);
+        if (!bookingGenerators.containsKey(state)) {
+            log.error("Unknown state: UNSUPPORTED_STATUS: {}", state);
+            throw new ValidationException("Unknown state: UNSUPPORTED_STATUS");
         }
-        if (state.equals("FUTURE")) {
-            List<Booking> bookings = bookingRepository.findAllByBooker(user);
-            log.info("Возвращаем список будущих бронирований для пользователя с id: {} {}", userId, bookings);
-            return bookings.stream()
-                    .sorted(this::compareFuture)
-                    .map(BookingConverter::toBookingResponseDto)
-                    .collect(Collectors.toList());
-        }
-        if (state.equals("WAITING")) {
-            Status status = Status.WAITING;
-            List<Booking> bookings = bookingRepository.findAllByBookerAndStatusEquals(user, status);
-            log.info("Возвращаем список ожидающих бронирований для пользователя с id: {} {}", userId, bookings);
-            return bookings.stream()
-                    .sorted(this::compareAll)
-                    .map(BookingConverter::toBookingResponseDto)
-                    .collect(Collectors.toList());
-        }
-        if (state.equals("REJECTED")) {
-            Status status = Status.REJECTED;
-            List<Booking> bookings = bookingRepository.findAllByBookerAndStatusEquals(user, status);
-            log.info("Возвращаем список отклоненных бронирований для пользователя с id: {} {}", userId, bookings);
-            return bookings.stream()
-                    .sorted(this::compareAll)
-                    .map(BookingConverter::toBookingResponseDto)
-                    .collect(Collectors.toList());
-        }
-        if (state.equals("CURRENT")) {
-            LocalDateTime now = LocalDateTime.now();
-            List<Booking> bookings = bookingRepository.findAllByBookerAndStartBeforeAndEndAfter(user, now, now);
-            log.info("Возвращаем список текущих бронирований для пользователя с id: {} {}", userId, bookings);
-            return bookings.stream()
-                    .map(BookingConverter::toBookingResponseDto)
-                    .collect(Collectors.toList());
-        }
-        if (state.equals("PAST")) {
-            LocalDateTime now = LocalDateTime.now();
-            List<Booking> bookings = bookingRepository.findAllByBookerAndEndBefore(user, now);
-            log.info("Возвращаем список прошедших бронирований для пользователя с id: {} {}", userId, bookings);
-            return bookings.stream()
-                    .map(BookingConverter::toBookingResponseDto)
-                    .collect(Collectors.toList());
-        }
-
-        log.error("Unknown state: UNSUPPORTED_STATUS: {}", state);
-        throw new ValidationException("Unknown state: UNSUPPORTED_STATUS");
+        return bookingUserGenerator.getUserBookings(userId);
     }
 
     @Override
     public List<BookingResponseDto> getOwnerBookings(String state, Long userId) {
-        User owner = userRepository.findById(userId).orElseThrow(() -> new NotFoundException("Пользователя с таким id" +
-                " не существует"));
-        if ("ALL".equals(state)) {
-            Item item = itemRepository.findDistinctTopByOwner(owner);
-            List<Booking> bookings = bookingRepository.findAllByItem(item);
-            log.info("Возвращаем список всех бронирований для владельца с id: {} {}", userId, bookings);
-            return bookings.stream()
-                    .sorted(this::compareAll)
-                    .map(BookingConverter::toBookingResponseDto)
-                    .collect(Collectors.toList());
+        BookingGenerator bookingOwnerGenerator = bookingGenerators.get(state);
+        if (!bookingGenerators.containsKey(state)) {
+            log.error("Unknown state: UNSUPPORTED_STATUS: {}", state);
+            throw new ValidationException("Unknown state: UNSUPPORTED_STATUS");
         }
-        if (state.equals("FUTURE")) {
-            Item item = itemRepository.findDistinctTopByOwner(owner);
-            List<Booking> bookings = bookingRepository.findAllByItem(item);
-            log.info("Возвращаем список будущих бронирований для владельца с id: {} {}", userId, bookings);
-            return bookings.stream()
-                    .sorted(this::compareFuture)
-                    .map(BookingConverter::toBookingResponseDto)
-                    .collect(Collectors.toList());
-        }
-        if (state.equals("WAITING")) {
-            List<Item> items = itemRepository.findByOwner(owner);
-            Status status = Status.WAITING;
-            List<Booking> bookings = new ArrayList<>();
-            for (Item item : items) {
-                if (item.getOwner().getId().equals(userId)) {
-                    Booking booking = bookingRepository.findByItemAndStatusEquals(item, status);
-                    if (booking != null) {
-                        bookings.add(booking);
-                        log.info("Возвращаем список отклонненых бронирований для владельца с id: {} {}", userId, bookings);
-                        return bookings.stream()
-                                .sorted(this::compareAll)
-                                .map(BookingConverter::toBookingResponseDto)
-                                .collect(Collectors.toList());
-                    }
-                }
-            }
-        }
-        if (state.equals("REJECTED")) {
-            List<Item> items = itemRepository.findByOwner(owner);
-            Status status = Status.REJECTED;
-            List<Booking> bookings = new ArrayList<>();
-            for (Item item : items) {
-                if (item.getOwner().getId().equals(userId)) {
-                    Booking booking = bookingRepository.findByItemAndStatusEquals(item, status);
-                    if (booking != null) {
-                        bookings.add(booking);
-                        log.info("Возвращаем список отклонненых бронирований для владельца с id: {} {}", userId, bookings);
-                        return bookings.stream()
-                                .sorted(this::compareAll)
-                                .map(BookingConverter::toBookingResponseDto)
-                                .collect(Collectors.toList());
-                    }
-                }
-            }
-        }
-        if (state.equals("CURRENT")) {
-            List<Item> items = itemRepository.findByOwner(owner);
-            LocalDateTime now = LocalDateTime.now();
-            List<Booking> bookings = new ArrayList<>();
-            for (Item item : items) {
-                if (item.getOwner().getId().equals(userId)) {
-                    Booking booking = bookingRepository.findByItemAndStartBeforeAndEndAfter(item, now, now);
-                    if (booking != null) {
-                        bookings.add(booking);
-                        log.info("Возвращаем список текущих бронирований для владельца с id: {} {}", userId, bookings);
-                        return bookings.stream()
-                                .map(BookingConverter::toBookingResponseDto)
-                                .collect(Collectors.toList());
-                    }
-                }
-            }
-        }
-        if (state.equals("PAST")) {
-            Item item = itemRepository.findDistinctTopByOwner(owner);
-            LocalDateTime now = LocalDateTime.now();
-            List<Booking> bookings = bookingRepository.findAllByItemAndEndBefore(item, now);
-            log.info("Возвращаем список прошедших бронирований для владельца с id: {} {}", userId, bookings);
-            return bookings.stream()
-                    .map(BookingConverter::toBookingResponseDto)
-                    .collect(Collectors.toList());
-        }
-        log.error("Unknown state: UNSUPPORTED_STATUS: {}", state);
-        throw new
-                ValidationException("Unknown state: UNSUPPORTED_STATUS");
-    }
-
-    private int compareAll(Booking booking1, Booking booking2) {
-        return booking2.getEnd().compareTo(booking1.getEnd());
-    }
-
-    private int compareFuture(Booking booking1, Booking booking2) {
-        LocalDateTime currentMoment = LocalDateTime.now();
-        return (booking2.getEnd().compareTo(currentMoment)) - (booking1.getEnd().compareTo(currentMoment));
+        return bookingOwnerGenerator.getOwnerBookings(userId);
     }
 }
