+++ src/main/java/ru/practicum/shareit/booking/Status.java
@@ -2,7 +2,6 @@ package ru.practicum.shareit.booking;
 
 public enum Status {
     WAITING,
-    CURRENT,
     APPROVED,
     REJECTED;
 }
